requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1
redis==5.0.1
oss2==2.18.4
boto3==1.34.0
openai==1.3.0
Pillow==10.1.0
requests==2.31.0
aiohttp==3.9.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-json-logger==2.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
tabulate==0.9.0
click==8.1.7
python-dotenv==1.0.0